Generate debug information for all MSVC builds.
We want to have debug information even for the release builds as it doesn't
affect the binaries produced but greatly helps when debugging, so create PDB
for all release builds as well in the manually maintained MSVS files (this was
already the case for Win32 release builds but not x64 ones for some reason).

This also fixes the problem with "None" being an invalid value for
DebugInformationFormat with MSVC 10 as a side effect.

Closes #16236.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@76475 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775