Update MSVC 10+ projects with upmake.
Update the files used in MSVC 10, 11 and 12 projects from the master files
list.

The main change is the removal of all headers from the core project: it
doesn't seem to make sense to have all headers, even for the completely
unrelated projects, in it and it is easier to just have the core own headers.

The only other changes are the addition of a few headers (including the .cpp
files used as headers) to the base project and the removal of the files which
are never built (but still appear in auto-generated by bakefile projects
because they could have been used in wxUniv case) in the core project.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@76611 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775